Fix for Commit after Rollback
In case something in `answers` has failed, the transaction has been
rolled back and put in a zombie state. When we Commit, we trigger an
exception. The throw would happen in the line after Commit (if we ever
got there).

By awaiting the response before Commit, the exception will be raised so
we don't try to commit in case of failure.